include header and uncle validation in consensus strategy
