+++ ethereum/ethpow_utils.py
@@ -10,6 +10,7 @@ import rlp
 from ethereum.state_transition import calc_difficulty, check_gaslimit, \
     initialize
 from ethereum.config import default_config
+from ethereum.exceptions import VerificationFailed
 
 # Gas limit adjustment algo
 def calc_gaslimit(parent, config=default_config):
@@ -59,3 +60,38 @@ def ethereum1_setup_block(chain, state=None, timestamp=None, coinbase='\x35'*20,
     blk.header.uncles_hash = sha3(rlp.encode(blk.uncles))
     initialize(state, blk)
     return blk
+
+
+def ethereum1_validate_header(state, header):
+    assert header.check_pow()
+    parent = state.prev_headers[0]
+    if parent:
+        if header.prevhash != parent.hash:
+            raise ValueError("Block's prevhash and parent's hash do not match: block prevhash %s parent hash %s" %
+                             (encode_hex(header.prevhash), encode_hex(parent.hash)))
+        if header.number != parent.number + 1:
+            raise ValueError("Block's number is not the successor of its parent number")
+        if not check_gaslimit(parent, header.gas_limit, config=state.config):
+            raise ValueError("Block's gaslimit is inconsistent with its parent's gaslimit")
+        if header.difficulty != calc_difficulty(parent, header.timestamp, config=state.config):
+            raise ValueError("Block's difficulty is inconsistent with its parent's difficulty: parent %d expected %d actual %d" %
+                             (parent.difficulty, calc_difficulty(parent, header.timestamp, config=state.config), header.difficulty))
+        if header.gas_used > header.gas_limit:
+            raise ValueError("Gas used exceeds gas limit")
+        if len(header.extra_data) > 32 and not state.is_SERENITY():
+            raise ValueError("Extra data too long")
+        if len(header.extra_data) > 1024:
+            raise ValueError("Extra data too long")
+        if header.timestamp <= parent.timestamp:
+            raise ValueError("Timestamp equal to or before parent")
+        if header.timestamp >= 2**256:
+            raise ValueError("Timestamp waaaaaaaaaaayy too large")
+    if header.gas_limit >= 2**63:
+        raise ValueError("Header gas limit too high")
+    return True
+
+def ethereum1_validate_uncle(state, uncle):
+    if not uncle.check_pow():
+        raise VerificationFailed('pow mismatch')
+    return True
+
